Fix more GLenum→int conversions: volumeslicegl.cpp OptionPropertyInt, ssao.cpp swizzle masks
- volumeslicegl.cpp: GL_CLAMP_TO_EDGE/GL_REPEAT/GL_MIRRORED_REPEAT in
  OptionPropertyInt need static_cast<int>()
- ssao.cpp: GLint swizzle[] arrays with GL_RED/GL_GREEN/GL_ZERO need
  static_cast<GLint>()

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/cd57c4bb-34b3-4732-bda3-04fa02e6f88a

+++ modules/postprocessing/src/processors/ssao.cpp
@@ -363,10 +363,12 @@ void SSAO::initFramebuffers(int width, int height) {
 
 #if USE_AO_SPECIAL_BLUR
     GLenum formatAO = GL_RG16F;
-    GLint swizzle[4] = {GL_RED, GL_GREEN, GL_ZERO, GL_ZERO};
+    GLint swizzle[4] = {static_cast<GLint>(GL_RED), static_cast<GLint>(GL_GREEN),
+                        static_cast<GLint>(GL_ZERO), static_cast<GLint>(GL_ZERO)};
 #else
     GLenum formatAO = GL_R8;
-    GLint swizzle[4] = {GL_RED, GL_RED, GL_RED, GL_RED};
+    GLint swizzle[4] = {static_cast<GLint>(GL_RED), static_cast<GLint>(GL_RED),
+                        static_cast<GLint>(GL_RED), static_cast<GLint>(GL_RED)};
 #endif
 
     newTexture(textures_.hbaoResult);
